(chore) rewrite to Operaton
This commits migrates a majority of the codebase from camunda
to operaton.
The new Version is based on the current Operaton version.
Packages and Classes targeting the Camunda Platform 8 were
removed.
Since we do not use all test classes and Maven modules
internally, not everything has been tested.
Some follow-up work will certainly be necessary, but
this can be considered a starting point.